--- project/src/components/app/app.tsx
@@ -1,13 +1,0 @@
-import MainScreen from '../main-screen/main-screen';
-
-type AppScreenProps = {
-  placesCount: number;
-}
-
-function App({placesCount}: AppScreenProps): JSX.Element {
-  return (
-    <MainScreen placesCount={placesCount} />
-  );
-}
-
-export default App;
