+++ project/src/components/navigation.tsx
@@ -1,0 +1,25 @@
+import {Link} from 'react-router-dom';
+import {AppRoute} from '../consts';
+
+function Navigation(): JSX.Element {
+  return (
+    <nav className="header__nav">
+      <ul className="header__nav-list">
+        <li className="header__nav-item user">
+          <Link to={AppRoute.Favorites} title='/favorites' className="header__nav-link header__nav-link--profile">
+            <div className="header__avatar-wrapper user__avatar-wrapper">
+            </div>
+            <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+          </Link>
+        </li>
+        <li className="header__nav-item">
+          <Link to={AppRoute.Login} title='/login' className="header__nav-link">
+            <span className="header__signout">Sign out</span>
+          </Link>
+        </li>
+      </ul>
+    </nav>
+  );
+}
+
+export default Navigation;
