+++ project/src/components/place-card.tsx
