+++ project/src/components/private-route.tsx
@@ -1,0 +1,15 @@
+import {Navigate} from 'react-router-dom';
+import {AuthorizationStatus} from '../consts';
+
+type PrivateRouteProps = {
+  authorizationStatus: AuthorizationStatus;
+  children: JSX.Element;
+};
+
+function PrivateRoute(props: PrivateRouteProps): JSX.Element {
+  const {authorizationStatus, children} = props;
+
+  return authorizationStatus === AuthorizationStatus.Auth ? children : <Navigate to={'/login'} />;
+}
+
+export default PrivateRoute;
