+++ project/src/consts.ts
@@ -1,0 +1,14 @@
+enum AppRoute {
+  Login = '/login',
+  Favorites = '/favorites',
+  Offer = '/offer/:id',
+  Root = '/'
+}
+
+enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN',
+}
+
+export {AppRoute, AuthorizationStatus};
