+++ project/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app/app';
+import App from './app/app';
 
 const Setting = {
   PLACES_COUNT: 500,
